refactor: merge backend, payload, runtime into reactive/client.py
Replace the generic _parse_client_payload higher-order function with
explicit MountedManifest/TriggerManifest/LoadedAssets parsers. Behavior
unchanged.
