+++ pyjinhx/core/base.py
@@ -189,8 +189,7 @@ class BaseComponent(BaseModel):
                 session=session,
             )
 
-        from pyjinhx.reactive.backend import ClientBackend
-        from pyjinhx.reactive.payload import LoadedAssets
+        from pyjinhx.reactive.client import ClientBackend, LoadedAssets
 
         resolved_client = ClientBackend.resolve_client(client)
         loaded_assets = (
