+++ pyjinhx/core/registry.py
@@ -142,7 +142,7 @@ class Registry:
         """
         from contextlib import ExitStack
 
-        from pyjinhx.reactive.backend import ClientBackend
+        from pyjinhx.reactive.client import ClientBackend
         from pyjinhx.reactive.load_cache import LoadCache
         from pyjinhx.reactive.context import LoadContext
         from pyjinhx.reactive.dev import warn_mutations_without_render
