+++ pyjinhx/reactive/dev.py
@@ -5,7 +5,7 @@ from dataclasses import dataclass
 
 from pyjinhx.core.registry import Registry
 
-from .backend import ClientBackend
+from .client import ClientBackend
 from .mutations import MutationTracker
 
 logger = logging.getLogger("pyjinhx")
